+++ game/232/AH_HintMenu.c
@@ -143,7 +143,7 @@ void AH_HintMenu_MenuProc(struct RectMenu *menu)
 		box.y = (s16)iVar11 + 0x4e;
 		box.h = 0x11;
 
-		u_long *ot = gGT->backBuffer->otMem.uiOT;
+		uint32_t *ot = gGT->backBuffer->otMem.uiOT;
 		CTR_Box_DrawClearBox(&box, &sdata->menuRowHighlight_Normal, TRANS_50_DECAL, ot);
 
 		box.y = 0x3c;
@@ -330,7 +330,7 @@ LAB_800b38cc:
 	box.y = (menu->rowSelected - D232.hintMenu_scrollIndex) * 0x10 + 0x4f;
 	box.h = 0x11;
 
-	u_long *ot = gGT->backBuffer->otMem.uiOT;
+	uint32_t *ot = gGT->backBuffer->otMem.uiOT;
 	CTR_Box_DrawClearBox(&box, &sdata->menuRowHighlight_Normal, TRANS_50_DECAL, ot);
 
 	box.y = 0x3c;
