+++ game/233/CS_Garage.c
@@ -212,7 +212,7 @@ void CS_Garage_MenuProc(struct RectMenu *param_1)
 				// pointer to OT memory
 				void *ot = gGT->pushBuffer_UI.ptrOT;
 
-				*(int *)p = CtrGpu_PackOTTag(*(u_long *)ot, 0x8000000);
+				*(int *)p = CtrGpu_PackOTTag(*(uint32_t *)ot, 0x8000000);
 				*(int *)ot = (int)CtrGpu_PrimToOTLink24(p);
 			}
