+++ game/CTR/CTR_Error.c
@@ -33,6 +33,6 @@ void CTR_ErrorScreen(char r, char g, char b)
 		p.g0 = g;
 		p.b0 = b;
 
-		DrawOTag((u_long *)&p);
+		DrawOTag((uint32_t *)&p);
 	}
 }
